# swagger.json
{"openapi":"3.0.0","paths":{"/api/users/signIn":{"post":{"operationId":"AppController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInResponseDto"}}}}}}},"/api/users/me":{"get":{"operationId":"AppController_getProfile","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"security":[{"bearer":[]}]}},"/api/users":{"post":{"operationId":"UsersController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["User"]},"get":{"operationId":"UsersController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}}},"tags":["User"],"security":[{"bearer":[]}]}},"/api/users/{id}":{"get":{"operationId":"UsersController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["User"],"security":[{"bearer":[]}]},"patch":{"operationId":"UsersController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["User"],"security":[{"bearer":[]}]}},"/api/products":{"post":{"operationId":"ProductsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProductDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Product"}}}}},"tags":["Product"],"security":[{"bearer":[]}]},"get":{"operationId":"ProductsController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Product"}}}}}},"tags":["Product"]}},"/api/products/{id}":{"get":{"operationId":"ProductsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Product"}}}}},"tags":["Product"]},"patch":{"operationId":"ProductsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProductDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Product"}}}}},"tags":["Product"],"security":[{"bearer":[]}]},"delete":{"operationId":"ProductsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Product"],"security":[{"bearer":[]}]}},"/api/orders":{"post":{"operationId":"OrdersController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateOrderDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Order"}}}}},"tags":["Order"],"security":[{"bearer":[]}]},"get":{"operationId":"OrdersController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Order"}}}}}},"tags":["Order"],"security":[{"bearer":[]}]}},"/api/orders/{id}":{"get":{"operationId":"OrdersController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Order"}}}}},"tags":["Order"],"security":[{"bearer":[]}]}}},"info":{"title":"ESOE DB","description":"The ESOE DB Api description","version":"1.0","contact":{}},"tags":[{"name":"ESOE","description":""}],"servers":[],"components":{"schemas":{"SignInDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"SignInResponseDto":{"type":"object","properties":{"id":{"type":"number"},"token":{"type":"string"}},"required":["id","token"]},"CreateUserDto":{"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"type":{"type":"string","enum":["買家","賣家"]},"phone":{"type":"string"}},"required":["name","email","password","type","phone"]},"User":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"},"type":{"enum":["買家","賣家"],"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"phone":{"type":"string"},"orders":{"type":"array","items":{"$ref":"#/components/schemas/Order"}}},"required":["id","name","type","email","password","phone","orders"]},"Product":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"},"description":{"type":"string"},"picture":{"type":"string"},"inventory":{"type":"number"},"price":{"type":"number"},"startSaleTime":{"format":"date-time","type":"string"},"endSaleTime":{"format":"date-time","type":"string"}},"required":["id","name","description","picture","inventory","price","startSaleTime","endSaleTime"]},"ProductOrder":{"type":"object","properties":{"orderId":{"type":"number"},"productId":{"type":"number"},"order":{"$ref":"#/components/schemas/Order"},"product":{"$ref":"#/components/schemas/Product"},"amount":{"type":"number"}},"required":["orderId","productId","order","product","amount"]},"Order":{"type":"object","properties":{"id":{"type":"number"},"buyerId":{"type":"number"},"buyer":{"$ref":"#/components/schemas/User"},"timestamp":{"format":"date-time","type":"string"},"productOrders":{"type":"array","items":{"$ref":"#/components/schemas/ProductOrder"}}},"required":["id","buyerId","buyer","timestamp","productOrders"]},"UpdateUserDto":{"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"type":{"type":"string","enum":["買家","賣家"]},"phone":{"type":"string"}}},"EsoeApiResponseDto":{"type":"object","properties":{"status":{"type":"number"},"message":{"type":"string"}},"required":["status","message"]},"CreateProductDto":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"picture":{"type":"string"},"inventory":{"type":"number"},"price":{"type":"number"},"startSaleTime":{"format":"date-time","type":"string"},"endSaleTime":{"format":"date-time","type":"string"}},"required":["name","description","picture","inventory","price","startSaleTime","endSaleTime"]},"UpdateProductDto":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"picture":{"type":"string"},"inventory":{"type":"number"},"price":{"type":"number"},"startSaleTime":{"format":"date-time","type":"string"},"endSaleTime":{"format":"date-time","type":"string"}}},"CreateProductOrderDto":{"type":"object","properties":{"productId":{"type":"number"},"amount":{"type":"number"}},"required":["productId","amount"]},"CreateOrderDto":{"type":"object","properties":{"productOrders":{"type":"array","items":{"$ref":"#/components/schemas/CreateProductOrderDto"}}},"required":["productOrders"]}}}}